+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 The changelog is based on the github's release or important commits.
 
-## v1.0.1 (28/01/2024)
+## v0.0.2 (31/01/2024)
+
+- Login and Signup page
+- Server-side login
+- Login Cookies
+- Auth tokens
+
+## v0.0.1 (28/01/2024)
 
 - Added Signup page and functionality of client and server (Exept cookie login)
 
@@ -11,7 +18,7 @@ The changelog is based on the github's release or important commits.
 - I need to reduce the amount of packages in the [python api](./api/server.py)
 - Needs cookies 🍪🍪🍪
 
-## v1.0.0 (27/01/2024)
+## v0.0.0 (27/01/2024)
 
 - Server (Python Flask),
 - API (`/api/user/add` and `/api/user/getfromid/`),